requirements.txt
pandas>=1.4.3
pyarrow>=9.0.0
aiofiles>=22.1.0
colorama>=0.4.5
aiohttp>=3.8.3
#open_clip_torch>=1.26.12
timm
fairscale==0.4.4
transformers==4.31.0
opencv-python>=4.6.0
pillow==10.0.0
termcolor==2.3.0
argparse
glob
os
cv2
shutil
pyexifinfo>=0.4.0
json
traceback